UPDATE: Adicionado sweetalert e reconfigurado os alerts dos templates

+++ webapp/assets/js/publicacoes.js
@@ -19,7 +19,11 @@ function criarPublicacao(evento){
     }).done(function(){
         window.location = "/home";
     }).fail(function(){
-        alert("Erro ao criar a publicação");
+        Swal.fire({
+            title: "Ops...!",
+            text: "Erro ao criar a publicação",
+            icon: "error"
+          });
     })
 
 }
@@ -45,7 +49,11 @@ function curtirPublicacao(evento) {
         elementoClicado.addClass('text-danger');
         elementoClicado.removeClass('curtir-publicacao');
     }).fail(function(){
-        alert("Erro ao curtir a publicação");
+        Swal.fire({
+            title: "Ops...!",
+            text: "Erro ao curtir a publicação",
+            icon: "error"
+        });
     }).always(function(){
         elementoClicado.prop('disabled', false);
     })
@@ -72,7 +80,11 @@ function descurtirPublicacao(evento){
         elementoClicado.removeClass('text-danger');
         elementoClicado.addClass('curtir-publicacao');
     }).fail(function(){
-        alert("Erro ao descurtir a publicação");
+        Swal.fire({
+            title: "Ops...!",
+            text: "Erro ao descurtir a publicação",
+            icon: "error"
+        });
     }).always(function(){
         elementoClicado.prop('disabled', false);
     })
@@ -92,9 +104,19 @@ function atualizarPublicacao(evento) {
             conteudo: $('#conteudo').val()
         }
     }).done(function(){
-        alert("Deu certo");
+        Swal.fire({
+            title: "Sucesso!",
+            text: "Publicação atualizada com sucesso",
+            icon: "success"
+          }).then(function(){
+            window.location = "/home";
+          })
     }).fail(function(){
-        alert("Deu fail");
+        Swal.fire({
+            title: "Ops...!",
+            text: "Falha em atualizar publicação",
+            icon: "error"
+        });
     }).always(function(){
         $('#atualizar-publicacao').prop('disable', false);
     })
@@ -102,22 +124,38 @@ function atualizarPublicacao(evento) {
 
 function deletarPublicacao(evento){
     evento.preventDefault();
+
+    Swal.fire({
+        title: "Atenção!",
+        text: "Tem certeza que deseja excluir a publicação? É uma ação irreversível!",
+        showCancelButton: true,
+        cancelButtonText: "Cancelar",
+        icon: "warning"
+
+      }).then(function(confirmacao){
+        if (!confirmacao.value) return;
+
+        const elementoClicado = $(evento.target);
+        const publicacao = elementoClicado.closest('div')
+        const publicacaoId = publicacao.data('publicacao-id');
     
-    const elementoClicado = $(evento.target);
-    const publicacao = elementoClicado.closest('div')
-    const publicacaoId = publicacao.data('publicacao-id');
-
-    elementoClicado.prop('disabled', true);
-    console.log("aqui");
-    $.ajax({
-        url: `publicacoes/${publicacaoId}`,
-        method: "DELETE"
-    }).done(function(){
-        publicacao.fadeOut("slow", function(){
-            $(this).remove();
-
+        elementoClicado.prop('disabled', true);
+        console.log("aqui");
+        $.ajax({
+            url: `publicacoes/${publicacaoId}`,
+            method: "DELETE"
+        }).done(function(){
+            publicacao.fadeOut("slow", function(){
+                $(this).remove();
+    
+            })
+        }).fail(function(){
+            Swal.fire({
+                title: "Ops...!",
+                text: "Erro ao excluir a publicação!",
+                icon: "error"
+            });
         })
-    }).fail(function(){
-        alert("Erro ao excluir a publicação!");
-    })
+      })
+    
 }
